add premissions handling to delete item mutation

+++ src/resolvers/Mutation.js
@@ -37,7 +37,13 @@ const Mutations = {
     },
     async deleteItem(parent, args, ctx, info) {
         const where = {id: args.id};
-        const item = await ctx.db.query.item({where}, `{ id title}`);
+        const item = await ctx.db.query.item({where}, `{ id title user {id }}`);
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+
+        if (!ownsItem || hasPermissions) {
+            throw new Error('You don\'t have permission to do that!')
+        }
         return ctx.db.mutation.deleteItem({where}, info);
     },
     async signup(parent, args, ctx, info) {
@@ -141,12 +147,14 @@ const Mutations = {
     },
 
     async updatePermissions(parent, args, ctx, info) {
-        if(!ctx.request.userId) {
+        if (!ctx.request.userId) {
             throw new Error('You must be logged in')
         }
-        const currentUser = await ctx.db.query.user({where: {
-            id: ctx.request.userId
-        }},
+        const currentUser = await ctx.db.query.user({
+                where: {
+                    id: ctx.request.userId
+                }
+            },
             info
         );
         hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE'])
